auto-research/requirements.txt
numpy
pandas
Pillow
scikit-learn
scipy
matplotlib
